backend-python/requirements.txt
uagents>=0.12.0
hyperon>=0.1.8
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
